refactor/cleanup packages
- delete the `version` package
- move the `tiers` and `wait` packages under `testsupport`
