+++ test/perf/perf_test.go
@@ -11,8 +11,8 @@ import (
 
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/pkg/apis/toolchain/v1alpha1"
 	. "github.com/codeready-toolchain/toolchain-e2e/testsupport"
-	"github.com/codeready-toolchain/toolchain-e2e/wait"
-	. "github.com/codeready-toolchain/toolchain-e2e/wait"
+	"github.com/codeready-toolchain/toolchain-e2e/testsupport/wait"
+	. "github.com/codeready-toolchain/toolchain-e2e/testsupport/wait"
 
 	"github.com/go-logr/logr"
 	framework "github.com/operator-framework/operator-sdk/pkg/test"
