+++ testsupport/tier_setup.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/pkg/apis/toolchain/v1alpha1"
-	. "github.com/codeready-toolchain/toolchain-e2e/wait"
+	. "github.com/codeready-toolchain/toolchain-e2e/testsupport/wait"
 
 	"github.com/operator-framework/operator-sdk/pkg/test"
 	corev1 "k8s.io/api/core/v1"
